+++ doc/CHANGELOG.md
@@ -49,6 +49,7 @@
 - `[Tree|Splitter]` Fixed horizontal scrollbar showing on tree-grid even after tree is collapsed and no longer in need of scrollbar. ([#1836](https://github.com/infor-design/enterprise-wc/issues/1836))
 - `[Tree]` Re-Fixed selected event returning incorrect node data after adding children through addNodes. ([#1851](https://github.com/infor-design/enterprise-wc/issues/1851))
 - `[LineChart]` Converted line chart tests to playwright. ([#1946](https://github.com/infor-design/enterprise-wc/issues/1946))
+- `[SwipeAction]` Converted swipe action tests to playwright. ([#1976](https://github.com/infor-design/enterprise-wc/issues/1976))
 
 ## 1.0.0-beta.22